--- src/op_analytics/datasources/chainsmeta/systemconfig/chainsystemconfig.py
@@ -1,215 +1,0 @@
-from dataclasses import dataclass
-from datetime import date
-
-import polars as pl
-
-from op_analytics.coreutils.logger import structlog
-from op_analytics.coreutils.misc import raise_for_schema_mismatch
-from op_analytics.coreutils.request import new_session
-
-from ..dataaccess import ChainsMeta
-from .systemconfig import SystemConfig
-
-log = structlog.get_logger()
-
-ETHEREUM_RPC_URL = "https://ethereum-rpc.publicnode.com"
-
-# Schema for system config data
-SYSTEM_CONFIG_SCHEMA = pl.Schema(
-    {
-        "name": pl.String,
-        "identifier": pl.String,
-        "chain_id": pl.Int64,
-        "rpc_url": pl.String,
-        "system_config_proxy": pl.String,
-        "block_number": pl.Int64,
-        "block_timestamp": pl.Int64,
-        "batch_inbox_slot": pl.String,
-        "dispute_game_factory_slot": pl.String,
-        "l1_cross_domain_messenger_slot": pl.String,
-        "l1_erc721_bridge_slot": pl.String,
-        "l1_standard_bridge_slot": pl.String,
-        "optimism_mintable_erc20_factory_slot": pl.String,
-        "optimism_portal_slot": pl.String,
-        "start_block_slot": pl.String,
-        "unsafe_block_signer_slot": pl.String,
-        "version": pl.Utf8,  # Changed to String to handle large 256-bit integers, nullable if not supported
-        "basefee_scalar": pl.Int64,
-        "batch_inbox": pl.Utf8,
-        "batcher_hash": pl.Utf8,
-        "blob_basefee_scalar": pl.Int64,
-        "dispute_game_factory": pl.Utf8,
-        "eip1559_denominator": pl.Int64,
-        "eip1559_elasticity": pl.Int64,
-        "gas_limit": pl.Int64,
-        "l1_cross_domain_messenger": pl.Utf8,
-        "l1_erc721_bridge": pl.Utf8,
-        "l1_standard_bridge": pl.Utf8,
-        "maximum_gas_limit": pl.Int64,
-        "minimum_gas_limit": pl.Int64,
-        "operator_fee_constant": pl.Int64,  # Null for chains that don't support this method
-        "operator_fee_scalar": pl.Int64,  # Null for chains that don't support this method
-        "optimism_mintable_erc20_factory": pl.Utf8,
-        "optimism_portal": pl.Utf8,
-        "overhead": pl.Utf8,  # Changed to String to handle large 256-bit integers, nullable if not supported
-        "owner": pl.Utf8,
-        "scalar": pl.Utf8,  # Changed to String to handle large 256-bit integers, nullable if not supported
-        "start_block": pl.Utf8,
-        "unsafe_block_signer": pl.Utf8,
-        "version_hex": pl.Utf8,
-        "dt": pl.Date,
-    }
-)
-
-DEFAULT_SPEED_BUMP = 0.4
-
-
-# --- Helpers ---
-def _normalize_address(val: str | None) -> str | None:
-    return val.lower() if isinstance(val, str) else val
-
-
-def _int_to_str(val):
-    return str(val) if val is not None else None
-
-
-@dataclass
-class ChainSystemConfig:
-    """A single system config for a blockchain."""
-
-    name: str
-    identifier: str
-    chain_id: int
-    rpc_url: str
-    system_config_proxy: str
-
-    def fetch(self, process_dt: date) -> dict | None:
-        """Call RPC and return system config data as a dict, or None on error."""
-        session = new_session()
-        system_config = SystemConfig(system_config_proxy=self.system_config_proxy)
-        config_metadata = system_config.call_rpc(
-            rpc_endpoint=ETHEREUM_RPC_URL,
-            session=session,
-            speed_bump=DEFAULT_SPEED_BUMP,
-        )
-        if config_metadata is None:
-            # an error was encountered
-            log.warning(f"error encountered for chain {self.identifier}")
-            return None
-        row = config_metadata.to_dict()
-        row.update(
-            {
-                "name": self.name,
-                "identifier": self.identifier,
-                "chain_id": self.chain_id,
-                "rpc_url": self.rpc_url,
-                "system_config_proxy": self.system_config_proxy,
-                "dt": process_dt,
-                # Convert large ints to str for polars
-                "scalar": _int_to_str(row.get("scalar")),
-                "overhead": _int_to_str(row.get("overhead")),
-                "version": _int_to_str(row.get("version")),
-            }
-        )
-        log.info(f"fetched system config metadata from rpc for chain {self.identifier}")
-        return row
-
-
-@dataclass
-class SystemConfigList:
-    """System config data."""
-
-    system_config_df: pl.DataFrame
-
-    @staticmethod
-    def store_system_config_data(system_config_data: list[dict]) -> "SystemConfigList | None":
-        """Store system config data to GCS and return SystemConfigList, or None if empty."""
-        if not system_config_data:
-            log.warning("No system config data to store")
-            return None
-        system_config_df = pl.DataFrame(system_config_data)
-        # Reorder columns
-        col_order = [
-            "name",
-            "identifier",
-            "chain_id",
-            "rpc_url",
-            "system_config_proxy",
-            "block_number",
-            "block_timestamp",
-            "batch_inbox_slot",
-            "dispute_game_factory_slot",
-            "l1_cross_domain_messenger_slot",
-            "l1_erc721_bridge_slot",
-            "l1_standard_bridge_slot",
-            "optimism_mintable_erc20_factory_slot",
-            "optimism_portal_slot",
-            "start_block_slot",
-            "unsafe_block_signer_slot",
-            "version",
-            "basefee_scalar",
-            "batch_inbox",
-            "batcher_hash",
-            "blob_basefee_scalar",
-            "dispute_game_factory",
-            "eip1559_denominator",
-            "eip1559_elasticity",
-            "gas_limit",
-            "l1_cross_domain_messenger",
-            "l1_erc721_bridge",
-            "l1_standard_bridge",
-            "maximum_gas_limit",
-            "minimum_gas_limit",
-            "operator_fee_constant",
-            "operator_fee_scalar",
-            "optimism_mintable_erc20_factory",
-            "optimism_portal",
-            "overhead",
-            "owner",
-            "scalar",
-            "start_block",
-            "unsafe_block_signer",
-            "version_hex",
-            "dt",
-        ]
-        system_config_df = system_config_df.select(
-            [c for c in col_order if c in system_config_df.columns]
-        )
-        # Normalize address columns
-        address_columns = [
-            c
-            for c in system_config_df.columns
-            if (
-                "address" in c.lower()
-                or c.endswith("_proxy")
-                or c
-                in [
-                    "batch_inbox",
-                    "dispute_game_factory",
-                    "l1_cross_domain_messenger",
-                    "l1_erc721_bridge",
-                    "l1_standard_bridge",
-                    "optimism_mintable_erc20_factory",
-                    "optimism_portal",
-                    "owner",
-                    "start_block",
-                    "unsafe_block_signer",
-                ]
-            )
-        ]
-        for col in address_columns:
-            system_config_df = system_config_df.with_columns(
-                pl.col(col).map_elements(_normalize_address, return_dtype=pl.String)
-            )
-        # Check schema
-        raise_for_schema_mismatch(
-            actual_schema=system_config_df.schema,
-            expected_schema=SYSTEM_CONFIG_SCHEMA,
-        )
-        ChainsMeta.SYSTEM_CONFIG_LIST.write(
-            dataframe=system_config_df,
-            sort_by=["chain_id"],
-        )
-        log.info(f"Stored {len(system_config_data)} system config records to GCS")
-
-        return SystemConfigList(system_config_df=system_config_df)
